Implement generic derivations for `QueryParams` and `QueryResults`
I moved `convertError` from `Database.MySQL.Simpe.QueryResults` into
`Database.MySQL.Simple.Result` to avoid a cyclic dependency between
the `Database.MySQL.Simpe.QueryResults` and
`Database.MySQL.Simpe.QueryResults.Generic` that both make use of said
function.

Note that this change requires GHC Haskell.  We may want to hide this
behind a preprocessor flag for that reason.

I don't quite know if it makes sense to derive instances for
`U1` (types with a single constructor and no fields).  Although it is
certainly helpful during development to not get a type error when
trying to derive instances for a type that you are in the process of
defining, like:

    data User = User

I have also noted a concern with my generic implementation for
`(:*:)`.  I'll reproduce it here:

I'm concerned about this implementation since it's biased
towards one side.  Meaning that it's expecting the chain of
`(:*:)` to lean to the right.  As an example of the problem
consider the following:

    > data T = T Int Int Int deriving (Generic)
    > from (T 1 2 3)
    M1 {unM1 = M1 {unM1 = M1 {unM1 = K1 {unK1 = 1}} :*: (M1 {unM1 = K1 {unK1 = 2}} :*: M1 {unM1 = K1 {unK1 = 3}})}}

If the result in stead had been like this (note the re-bracketing):

    M1 {unM1 = (M1 {unM1 = M1 {unM1 = K1 {unK1 = 1}} :*: M1 {unM1 = K1 {unK1 = 2}}}) :*: M1 {unM1 = K1 {unK1 = 3}}}

Then the generic derivation for `T` would fail.

Furthermore, to be able to supply the arity in the call to
`convertError` I think we might need a `Typeable (a :*: b)`
constraint.

Lastly, the existing implementations for `QueryResults` uses
strictness annotations in certain places.  I don't know if these are
strictly (pun!) necessary and haven't used them in the generic
implementation.